Performance improvements, mostly to EditsMulti
It was doing a bunch of unnecessary work, and returning some strings
multiple times; now it's not!  Arguably this is a backwards
compatibility break; and the implementation is certainly more complex,
so it may or may not be desirable.

The added benchmark is now way faster:
```
before             92659             12744 ns/op           12736 B/op        139 allocs/op
after             628777              2122 ns/op            1616 B/op         39 allocs/op
```

I made a few other improvements where I noticed them.